[apps/shared] Templatize FloatParameterController to handle float/double parameters.
Fix bug: when entering "e^234" as a parameter of a model keeping floats,
the FloatParameterController would accept the number (because e^234 is
defined in double) and store an undefined value in the model (because
e^234 is undefined in float).

+++ apps/probability/parameters_controller.cpp
@@ -137,11 +137,11 @@ int ParametersController::reusableParameterCellCount(int type) {
   return m_distribution->numberOfParameter();
 }
 
-double ParametersController::parameterAtIndex(int index) {
+float ParametersController::parameterAtIndex(int index) {
   return m_distribution->parameterValueAtIndex(index);
 }
 
-bool ParametersController::setParameterAtIndex(int parameterIndex, double f) {
+bool ParametersController::setParameterAtIndex(int parameterIndex, float f) {
   if (!m_distribution->authorizedValueAtIndex(f, parameterIndex)) {
     Container::activeApp()->displayWarning(I18n::Message::ForbiddenValue);
     return false;
